Tear down and rebuild routes on reconnect call.
This is required if we roam from networkA to networkB, and networkB has a
different subnet than networkA. The route to $VPNGATEWAY will have to change.
We cannot simply tear down and rebuild the $VPNGATEWAY route, because if we
do, $VPNGATEWAY may end up routing over the tunnel.

TODO: smarter $VPNGATEWAY route detection so we can only tear down and bring
up that route.

Signed-off-by: Alexander Huynh <alex@frothy.coffee>